requirements.txt
albumentations==1.0.3
backcall==0.2.0
braceexpand==0.1.7
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.4
cloudpickle==2.0.0
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
cytoolz==0.11.2
dask==2022.2.1
debugpy==1.3.0
decorator==5.0.9
fonttools==4.30.0
fsspec==2022.2.0
idna==3.3
imagecodecs-lite==2019.12.3
imageio==2.16.1
imgaug==0.4.0
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==0.17.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.2
locket==0.2.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
networkx==2.7.1
numpy==1.21.5
packaging==21.3
pandas==1.4.1
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.2
prompt-toolkit==3.0.19
pycparser==2.21
Pygments==2.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.1.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==0.23.2
scipy==1.7.3
setuptools==58.0.4
Shapely==1.8.0
six==1.16.0
threadpoolctl==2.1.0
tifffile==2019.7.26.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
unicodedata2==14.0.0
urllib3==1.26.8
wcwidth==0.2.5
webdataset==0.1.103
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
